requirements.txt
pyyaml
scrapy
scrapy-splash
bs4
django
nltk
numpy
urllib3[secure]
